items.id=276: fix Tauri CI -- fmt drift + clippy errors
cargo fmt --check and cargo clippy -- -D warnings (CI's exact commands)
have failed on every push since 2026-08-03. Reformats fmt drift across
the whole src-tauri tree (spread well past the 3 files first flagged),
switches pane_host.rs off the deprecated webkit2gtk run_javascript to
evaluate_javascript (v2_40 feature), drops an unnecessary cast in
tier3_pane.rs, fixes a doc-comment paragraph break in gl_loader.rs,
groups apply_synced_document's/apply_synced_document_conn's
document-payload params into a SyncedDocumentFields struct in
group_store.rs to clear too_many_arguments, and fixes a
manual_is_multiple_of in main.rs that only surfaced once the lib
target started compiling clean.

+++ src-tauri/src/commands/tier3_pane.rs
@@ -204,7 +204,7 @@ fn cef_cookie_to_stored(c: &Cookie) -> StoredCookie {
         secure: c.secure != 0,
         httponly: c.httponly != 0,
         same_site: c.same_site.get_raw() as i32,
-        priority: c.priority.get_raw() as i32,
+        priority: c.priority.get_raw(),
         has_expires,
         expires: if has_expires {
             Some(c.expires.val)
